Code cleanup related to drawing scanlines and testing graphics

+++ cpu.h
@@ -15,12 +15,11 @@ public:
     void set_PC(uint16_t bytes);
     void set_SP(uint16_t bytes);
 
-    int write_memory(uint8_t address, uint8_t data);
-    uint8_t read_memory(uint8_t address);
+    int write_memory(uint16_t address, uint8_t data);
+    uint8_t read_memory(uint16_t address);
 
 private:
     uint8_t memory[0x10000];
     uint16_t PC;
     uint16_t SP;
 };
-
